+++ Tests/MapboxDirectionsTests/VisualInstructionComponentTests.swift
@@ -23,10 +23,16 @@ class VisualInstructionComponentTests: XCTestCase {
     }
     
     func testImageComponent() {
-        let componentJSON = [
-            "text": "US 42",
+        let componentJSON: [String : Any] = [
+            "text": "I 95",
             "type": "icon",
-            "imageBaseURL": "https://s3.amazonaws.com/mapbox/shields/v3/us-42",
+            "imageBaseURL": "https://s3.amazonaws.com/mapbox/shields/v3/i-95",
+            "mapbox_shield": [
+                "base_url": "https://api.mapbox.com/styles/v1/",
+                "name": "us-interstate",
+                "text_color": "white",
+                "display_ref": "95"
+            ]
         ]
         let componentData = try! JSONSerialization.data(withJSONObject: componentJSON, options: [])
         var component: VisualInstruction.Component?
@@ -34,21 +40,109 @@ class VisualInstructionComponentTests: XCTestCase {
         XCTAssertNotNil(component)
         if let component = component {
             switch component {
-            case .image(let image, let alternativeText):
-                XCTAssertEqual(image.imageBaseURL?.absoluteString, "https://s3.amazonaws.com/mapbox/shields/v3/us-42")
-                XCTAssertEqual(image.imageURL(scale: 1, format: .svg)?.absoluteString, "https://s3.amazonaws.com/mapbox/shields/v3/us-42@1x.svg")
-                XCTAssertEqual(image.imageURL(scale: 3, format: .svg)?.absoluteString, "https://s3.amazonaws.com/mapbox/shields/v3/us-42@3x.svg")
-                XCTAssertEqual(image.imageURL(scale: 3, format: .png)?.absoluteString, "https://s3.amazonaws.com/mapbox/shields/v3/us-42@3x.png")
-                XCTAssertEqual(alternativeText.text, "US 42")
+            case .image(let image, let alternativeText, let shield):
+                XCTAssertEqual(image.imageBaseURL?.absoluteString, "https://s3.amazonaws.com/mapbox/shields/v3/i-95")
+                XCTAssertEqual(image.imageURL(scale: 1, format: .svg)?.absoluteString, "https://s3.amazonaws.com/mapbox/shields/v3/i-95@1x.svg")
+                XCTAssertEqual(image.imageURL(scale: 3, format: .svg)?.absoluteString, "https://s3.amazonaws.com/mapbox/shields/v3/i-95@3x.svg")
+                XCTAssertEqual(image.imageURL(scale: 3, format: .png)?.absoluteString, "https://s3.amazonaws.com/mapbox/shields/v3/i-95@3x.png")
+                XCTAssertEqual(alternativeText.text, "I 95")
                 XCTAssertNil(alternativeText.abbreviation)
                 XCTAssertNil(alternativeText.abbreviationPriority)
+                XCTAssertEqual(shield?.baseURL, URL(string: "https://api.mapbox.com/styles/v1/")!)
+                XCTAssertEqual(shield?.name, "us-interstate")
+                XCTAssertEqual(shield?.textColor, "white")
+                XCTAssertEqual(shield?.text, "95")
             default:
                 XCTFail("Image component should not be decoded as any other kind of component.")
             }
         }
         
-        component = .image(image: .init(imageBaseURL: URL(string: "https://s3.amazonaws.com/mapbox/shields/v3/us-42")!),
-                           alternativeText: .init(text: "US 42", abbreviation: nil, abbreviationPriority: nil))
+        component = .image(image: .init(imageBaseURL: URL(string: "https://s3.amazonaws.com/mapbox/shields/v3/i-95")!),
+                           alternativeText: .init(text: "I 95", abbreviation: nil, abbreviationPriority: nil),
+                           shield: .init(baseURL: URL(string: "https://api.mapbox.com/styles/v1/")!, name: "us-interstate", textColor: "white", text: "95"))
+        let encoder = JSONEncoder()
+        var encodedData: Data?
+        XCTAssertNoThrow(encodedData = try encoder.encode(component))
+        XCTAssertNotNil(encodedData)
+        
+        if let encodedData = encodedData {
+            var encodedComponentJSON: [String: Any?]?
+            XCTAssertNoThrow(encodedComponentJSON = try JSONSerialization.jsonObject(with: encodedData, options: []) as? [String: Any?])
+            XCTAssertNotNil(encodedComponentJSON)
+            
+            XCTAssert(JSONSerialization.objectsAreEqual(componentJSON, encodedComponentJSON, approximate: false))
+        }
+    }
+    
+    func testShield() {
+        let shieldJSON = [
+            "base_url": "https://api.mapbox.com/styles/v1/",
+            "name": "us-interstate",
+            "text_color": "white",
+            "display_ref": "95",
+        ]
+        
+        let shieldData = try! JSONSerialization.data(withJSONObject: shieldJSON, options: [])
+        var shield: VisualInstruction.Component.ShieldRepresentation?
+        XCTAssertNoThrow(shield = try JSONDecoder().decode(VisualInstruction.Component.ShieldRepresentation.self, from: shieldData))
+        XCTAssertNotNil(shield)
+        let url = URL(string: "https://api.mapbox.com/styles/v1/")
+        if let shield = shield {
+            XCTAssertEqual(shield.baseURL, url)
+            XCTAssertEqual(shield.name, "us-interstate")
+            XCTAssertEqual(shield.textColor, "white")
+            XCTAssertEqual(shield.text, "95")
+        }
+        shield = .init(baseURL: url!, name: "us-interstate", textColor: "white", text: "95")
+        
+        let encoder = JSONEncoder()
+        var encodedData: Data?
+        XCTAssertNoThrow(encodedData = try encoder.encode(shield))
+        XCTAssertNotNil(encodedData)
+        
+        if let encodedData = encodedData {
+            var encodedShieldJSON: [String: Any]?
+            XCTAssertNoThrow(encodedShieldJSON = try JSONSerialization.jsonObject(with: encodedData, options: []) as? [String: Any])
+            XCTAssertNotNil(encodedShieldJSON)
+            
+            XCTAssert(JSONSerialization.objectsAreEqual(shieldJSON, encodedShieldJSON, approximate: false))
+        }
+    }
+    
+    func testShieldImageComponent() {
+        let componentJSON: [String : Any] = [
+            "text": "I 95",
+            "type": "icon",
+            "mapbox_shield": [
+                "base_url": "https://api.mapbox.com/styles/v1/",
+                "name": "us-interstate",
+                "text_color": "white",
+                "display_ref": "95"
+            ]
+        ]
+        let componentData = try! JSONSerialization.data(withJSONObject: componentJSON, options: [])
+        var component: VisualInstruction.Component?
+        XCTAssertNoThrow(component = try JSONDecoder().decode(VisualInstruction.Component.self, from: componentData))
+        XCTAssertNotNil(component)
+        if let component = component {
+            switch component {
+            case .image(let image, let alternativeText, let shield):
+                XCTAssertNil(image.imageBaseURL?.absoluteString)
+                XCTAssertEqual(alternativeText.text, "I 95")
+                XCTAssertNil(alternativeText.abbreviation)
+                XCTAssertNil(alternativeText.abbreviationPriority)
+                XCTAssertEqual(shield?.baseURL, URL(string: "https://api.mapbox.com/styles/v1/")!)
+                XCTAssertEqual(shield?.name, "us-interstate")
+                XCTAssertEqual(shield?.textColor, "white")
+                XCTAssertEqual(shield?.text, "95")
+            default:
+                XCTFail("Image component should not be decoded as any other kind of component.")
+            }
+        }
+        
+        component = .image(image: .init(imageBaseURL: nil),
+                           alternativeText: .init(text: "I 95", abbreviation: nil, abbreviationPriority: nil),
+                           shield: .init(baseURL: URL(string: "https://api.mapbox.com/styles/v1/")!, name: "us-interstate", textColor: "white", text: "95"))
         let encoder = JSONEncoder()
         var encodedData: Data?
         XCTAssertNoThrow(encodedData = try encoder.encode(component))